BinUtils廃止。Float→固定小数点変換とFloatConverterにまとめた IAudioInputIteratorのTypoを修正。

+++ windows/TBSKmodem/getstarted/step3_bytedata/Program.cs
@@ -5,7 +5,7 @@ using jp.nyatla.kokolink.utils.wavefile;
 
 var tone = TbskTone.CreateXPskSin(10, 10).Mul(0.5);//    # SSFM DPSK
 var payload = Encoding.ASCII.GetBytes("0123456789");// # 10byte
-uint carrier = 8000;
+int carrier = 8000;
 
 //# modulation
 var mod = new TbskModulator(tone);
@@ -15,7 +15,7 @@ IList<double> src_pcm = mod.Modulate(payload).ToList();
 var wav = new PcmData(src_pcm, 16, carrier);
 using (var stream = File.Open("./step3.cs.wav", FileMode.Create, FileAccess.Write))
 {
-    var pcm = new PcmData(src_pcm, 16, (uint)carrier);
+    var pcm = new PcmData(src_pcm, 16, carrier);
     PcmData.Dump(pcm, stream);
 }
